+++ HISTORY.rst
@@ -19,6 +19,8 @@ unreleased
 * The command `az iot device-update device class list` adds support for `--filter` when no `--group-id` is provided.
 * The parameters `--account`, `--instance`, and `--resource-group` support setting default overridable values via config.
   Use `az config set` i.e. `az config set defaults.adu_account=<name>` or `az configure` i.e. `az configure --defaults adu_account=<name>`.
+* Introducing the experimental command `az iot device-update update init v5` for initializing (or generating) an import manifest
+  with the desired state.
 * Improved built-in documentation.